Change template to Material 3
