+++ core/ui/compose/designsystem/src/debug/kotlin/app/k9mail/core/ui/compose/designsystem/template/ResponsiveContentWithSurfacePreview.kt
@@ -5,36 +5,15 @@ import androidx.compose.foundation.layout.padding
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import app.k9mail.core.ui.compose.common.annotation.PreviewDevices
-import app.k9mail.core.ui.compose.designsystem.atom.Background
+import app.k9mail.core.ui.compose.designsystem.PreviewWithTheme
 import app.k9mail.core.ui.compose.designsystem.atom.Surface
-import app.k9mail.core.ui.compose.theme.K9Theme
-import app.k9mail.core.ui.compose.theme.MainTheme
-
-/**
- * The [ResponsiveContentWithBackground] composable embeds its content in [ResponsiveContent] with [Background].
- *
- * @param modifier The modifier to be applied to the layout.
- * @param content The content to be displayed.
- */
-@Composable
-fun ResponsiveContentWithBackground(
-    modifier: Modifier = Modifier,
-    content: @Composable () -> Unit,
-) {
-    ResponsiveContent {
-        Background(
-            modifier = modifier,
-        ) {
-            content()
-        }
-    }
-}
+import app.k9mail.core.ui.compose.theme2.MainTheme
 
 @Composable
 @PreviewDevices
 internal fun ResponsiveContentWithBackgroundPreview() {
-    K9Theme {
-        ResponsiveContentWithBackground {
+    PreviewWithTheme {
+        ResponsiveContentWithSurface {
             Surface(
                 color = MainTheme.colors.info,
                 modifier = Modifier
