+++ core/ui/compose/designsystem/src/main/kotlin/app/k9mail/core/ui/compose/designsystem/template/ResponsiveWidthContainer.kt
@@ -1,24 +1,17 @@
 package app.k9mail.core.ui.compose.designsystem.template
 
 import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.requiredWidth
-import androidx.compose.material.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
-import app.k9mail.core.ui.compose.common.annotation.PreviewDevices
 import app.k9mail.core.ui.compose.common.window.WindowSizeClass
 import app.k9mail.core.ui.compose.common.window.WindowSizeClass.Compact
 import app.k9mail.core.ui.compose.common.window.WindowSizeClass.Expanded
 import app.k9mail.core.ui.compose.common.window.WindowSizeClass.Medium
 import app.k9mail.core.ui.compose.common.window.getWindowSizeInfo
-import app.k9mail.core.ui.compose.designsystem.atom.Background
-import app.k9mail.core.ui.compose.designsystem.atom.Surface
-import app.k9mail.core.ui.compose.theme.K9Theme
-import app.k9mail.core.ui.compose.theme.MainTheme
 
 /**
  * A container that adjusts its width depending on the screen size.
@@ -64,40 +57,6 @@ fun ResponsiveWidthContainer(
             modifier = responsiveModifier,
         ) {
             content()
-        }
-    }
-}
-
-@Composable
-@PreviewDevices
-internal fun ResponsiveWidthContainerK9Preview() {
-    K9Theme {
-        Background {
-            ResponsiveWidthContainer {
-                Surface(
-                    color = MainTheme.colors.error,
-                    modifier = Modifier.fillMaxSize(),
-                ) {
-                    Text("Hello, World!")
-                }
-            }
-        }
-    }
-}
-
-@Composable
-@PreviewDevices
-internal fun ResponsiveWidthContainerThunderbirdPreview() {
-    K9Theme {
-        Background {
-            ResponsiveWidthContainer {
-                Surface(
-                    color = MainTheme.colors.error,
-                    modifier = Modifier.fillMaxSize(),
-                ) {
-                    Text("Hello, World!")
-                }
-            }
         }
     }
 }
